refactor: split mods_screen into components, dialogs and utils
Break the 4578-line mods_screen.dart down to ~1760 lines (behaviour
unchanged), extracting 14 focused files:

- utils: markdown_editor, markdown_description, url_utils; models: edit_image
- components: mods_action_buttons, mods_empty_states, mods_grouped_view,
  own_scroll_controller
- dialogs: rename, delete, edit, mod_details, keybinds, mod_context_menu

Dialogs follow the existing show*/callback pattern so reload timing is
preserved; the screen keeps the core stateful orchestration and per-mod
card wiring. Addresses BUGS & TODO #9.
